day_05_document_qa/requirements.txt
# Day 05 — Document Q&A with Vector Database
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ML / NLP
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0
chardet>=5.2.0

# Utilities
loguru>=0.7.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0

# Frontend
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0

# Data exploration
jupyter>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
